test list manager: better error message
We tried to protect Goofy by adding `CatchException` decorator to some
functions in test list manager.  The exception will be
logged and discarded, and the program is very likely to break in other
places, with a not very helpful exception (e.g. "None has no attribute
foo").  Since we have `NoRecursion` decorator which prevents infinite
recursive call when the test list is not resolvable, we should be able
to remove `CatchException` decorator and let goofy handle the exception.

BUG=b:31678190
TEST=make test, start goofy

Change-Id: I2e8c36eac622b8ba4ece8c3be46e00bfc95fe87c
Reviewed-on: https://chromium-review.googlesource.com/642638
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>