goofy: Change plugin 'system_log_manager' to use sever_proxy.
The system_log_manager is talking to factory server instead of
shopfloor backends. Moreover, the timeout for "getting instance" is not
really important because it should be using the global default settings
(usually we want to override the timeout for particular special
functions, especially SaveAuxLog).

BUG=chromium:760017
TEST=make test

Change-Id: I11dab30ca8d07e36b0d0be7c26625d1db556ca7c
Reviewed-on: https://chromium-review.googlesource.com/644829
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>