goofy.js: Refactor tab UI.
Refactor tab UI and extract everything that are related to another file,
so it's easier to add support for other UI layout.

Also refactor the logic of visible test. The old logic was that goofy
backend is the root of trust for which test is visible, and there's only
one visible test at a time. The new logic is that the UI manager is the
root of trust for which test is visible, and there may be multiple tests
that are visible at a time. (For preparation of upcoming tiled UI
feature.)

BUG=b:64963909
TEST=make test
TEST=manually on DUT.

Change-Id: If078ebf838ccec36034e4cf511fe20977dd39979
Reviewed-on: https://chromium-review.googlesource.com/647249
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>