cros_payload: Always update version information.
Version is now a core attribute that we have to set for every
components, even if the file is empty.

Without this, version will be wrong if we add a valid resource then
replace it with an empty file - the version string won't be changed.

BUG=chromium:761418
TEST=./cros_payload add test.json firmware empty_file

Change-Id: I0ff1e97d67cfefc8aae053a9286af0afdb6d3b98
Reviewed-on: https://chromium-review.googlesource.com/647731
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@chromium.org>