cros_payload: Cache sudo session earlier.
When adding an image resource, we usually have to mount it for fetching
version information that needs root permission.

Without this, we'll get the TODO prompt very late (after the compression
or first partition is done). As a result, we should cache sudo session
when an add_image command is found.

BUG=None
TEST=./cros_payload add test.json release_image PATH_TO_IMAGE

Change-Id: I9659372bcbdeb6d9fb0ff8dbaf06bafe2a4b3d31
Reviewed-on: https://chromium-review.googlesource.com/647730
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@chromium.org>