test: Move 'leds' to utils/kbd_leds.
The 'leds' in test top level folder may be misleading. It is actually a
module with utilities to blink LED on traditional external keyboards, as
a way to indicate device status when we can't reach the screen (for
example, test inside a chamber).

Changed module to cros.factory.test.utils.kbd_leds so the 'test' folder
contains only core modules.

BUG=None
TEST=make test

Change-Id: I5a02b16f1369286a41e179c729fd95d70be456c9
Reviewed-on: https://chromium-review.googlesource.com/647489
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>