overlord: fix goroutine leaks
There is a goroutine leakage in RPCCore.SpawnReaderRoutine method. The
cause of the leakage is because we send the data buffer to read channel
before checking error. Also since the read channel is unbuffered, it
always blocks. This caused the reader goroutine to be left indefinitely
in the memory. This CL fix the problem by making the read channel
buffered, and always check for error before sending data to the channel.

BUG=none
TEST=`overlordd`, launch 1000 ghost clients with:
     `for i in $(seq 1000); do ghost --mid Test --no-rpc-server
      --no-lan-disc &; done`
      Then terminate all go routines: `pkill -9 -f ghost`. The number of
      goroutines in the system should rename the same as before.

Change-Id: Iaeb52c370100f0959bcb591ecf13978958a9bbd2
Reviewed-on: https://chromium-review.googlesource.com/657457
Commit-Ready: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>