update_utils: Correct the check update process of HWID database.
If the HWID bundle on the factory server contains multiple HWID
databases, the version information of the HWID bundle will be a list
of checksum of each HWID database.  So instead of literally comparing
the checksum of the project specific HWID database on the DUT, we
should check whether the list of checksums contains local checksum
or not.

BUG=chromium:764616
TEST=manually test on DUT

Change-Id: I8076e07d33af01d185bbbc11960eae2164aa4e7d
Reviewed-on: https://chromium-review.googlesource.com/664617
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Youcheng Syu <youcheng@chromium.org>