utils: Add `GetDict` to retrieve config style dict values easier.
For configuration data, we usually have nested mapping structure and
reading values from them may be painful because we need to check
KeyError for every level, So we want to define a new getter function
that handles all non-exist intermediate mappings.

The new GetDict accepts a 'key_path', which can be a list of key names
(for example ['a', 'b'] or a simple key path delimited by dot (like "a.b").

BUG=None
TEST=make test

Change-Id: I5a2682a301ec653045731539a88e1c5912f77b6a
Reviewed-on: https://chromium-review.googlesource.com/666857
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@chromium.org>