config_utils: Remove the sepcial field "inherit" from resolved config.
The field "inherit" on the top-level of the JSON config dictionary
is reserved for listing dependent configs.  This CL modifies
`config_utils.LoadConfig` so that it removes this special field
before doing the schema check.  As the result, the json schema of the
config no longer need to define the format of "inherit" field.

BUG=None
TEST=manually test; make test

Change-Id: Id2ad27e338e3ae18dce74eb38597e2b85c69a547
Reviewed-on: https://chromium-review.googlesource.com/670205
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>