pytest: Fix camera e2e mode fail with InvalidStateError.
The MediaStreamTrack sometimes become muted when the system is busy,
which cause the e2e grabFrame to fail. Reinitialize the stream seems to
solve the problem.

Uploading image using event is quite slow, and we should add other ways
to upload binary file to Goofy backend in the future.

BUG=b:65388920
TEST=make test
TEST=manually on DUT.

Change-Id: I6b2ab55e179bbf842225c32a47c230c310a7b3b3
Reviewed-on: https://chromium-review.googlesource.com/672171
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>