Dome: expand the new bundle when updating
When updating a resource in a bundle, a new bundle card will be inserted
on the top of the bundle list. But it's by default collapsed. Users may
not notice the change and keep updating other resources in the same old
bundle. Say, if the user updated resource A, B, and C. That will create
3 new bundles but none of them contains all 3 updated changes (each of
them contains only one change from either A, B, or C).

To improve user experience, this CL collapse the old bundle and expand
the new bundle when updating a resource. That way, the user is more
likely to notice the change.

BUG=None
TEST=Manually tested

Change-Id: If432b0a12a76a573b433b44422fbc10f38b3c8ee
Reviewed-on: https://chromium-review.googlesource.com/674490
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>