goofy: Initialize logging in main.
The init_logging can be done only one time, but Goofy needs to parse
command line arguments first before deciding if it should init logging
with which verbosity. This creates a dirty hack in Goofy and its
unit test.

A better solution is to always do logging initialize in main. To do
that, we moved the command line parser into a new
GetCommandLineArgsParser method. In future we may consider moving this
to global function and let Goofy.init accepting only parser options and
arguments.

BUG=None
TEST=make test

Change-Id: I2f576178b95b1feea264c1e05fe2609b680ebf65
Reviewed-on: https://chromium-review.googlesource.com/676885
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>