test.css: Remove test-vcenter-{inner, outer}.
CSS flex layout is a more simple and less hacky solution for centering
things, change all tests to use that instead.

Also fix some pytest that use uninitialized variables.

Since the light_sensor_calibration is refactored in CL:547398 (and
test-vcenter-{inner, outer} is removed in that CL), and the test doesn't
work now, we leave it as is in this CL.

BUG=b:64963909
TEST=make test
TEST=manually on DUT.

Change-Id: I673a52044f0b2dfcc50ccc2e9641808a7561fb8b
Reviewed-on: https://chromium-review.googlesource.com/684102
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>