server: Reduce and suppress server connection error messages.
The TestLog captures all logging.* as individual events, and we've found
that many DUTs were not connected to network when the
`sync_factory_server` or `shopfloor_service` started. This will make
TestLog filled with useless "connection failure" events.
So both pytests are changed to only print same message one time.

Also, for most of the time, we think server proxy should not print
connection failure; instead it should be 'quiet' by default and let its
caller do the logging; so the `quiet` for ServerProxy is now True.

BUG=b:66870329
TEST=make test

Change-Id: I0a71ff76e5cc85e72ff795ea1349a1de6157672a
Reviewed-on: https://chromium-review.googlesource.com/686654
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>