server_proxy: Reset and retry when failed loading server config.
Tests writing runtime config may fail due to corrupted config files.
To recover from that, we added a DeleteRuntimeConfig to wipe configs in
runtime. Note we are not deleting configs in board config folder or
module default folder because those are supposed to be read only
and no way to recover.

`sync_factory_server` now will try to validate config in the very
beginning, and delete the runtime config if failed loading config.

Also, revised _LoadJsonFile so the path of corrupted config file is
logged.

BUG=b:67021050
TEST=make test;
     touch /var/factory/config/factory_server.json
     # run sync_factory_server.py and success.

Change-Id: I20ad39928b421514a5e0ce5e527f2ead3151170a
Reviewed-on: https://chromium-review.googlesource.com/689896
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>