Umpire: Change config format from YAML to JSON.
With the online Umpire config editor in Dome, we don't need to manually
edit Umpire config file in text editor anymore. It's time to change the
config format from YAML to JSON now.

This CL also changes UmpireConfig.__init__ to explicitly specify if the
input argument is a file path or not.

BUG=b:64731419,b:67252008
TEST=manually

Change-Id: I7fc93941ac4201ba8d491fff8dbc5149de777f18
Reviewed-on: https://chromium-review.googlesource.com/694766
Commit-Ready: Youcheng Syu <youcheng@chromium.org>
Tested-by: Youcheng Syu <youcheng@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>