test: Remove event_queue argument from EventClient class
The 'event_queue', along with the 'callback' parameter, are kinda
confusing at the first glance. The original design is:

1. If the 'callback' is specified, and 'event_queue' is not specified,
   then call 'callback' when an event arrives.
2. If the 'callback' is specified, and 'event_queue' is a Queue, then
   put a lambda of call-the-callback-with-arriving-event to the queue.
   In this case, the 'callback' should not be called upon event arrival.
3. If the 'callback' is not specified, do nothing when an event arrives.

It is found that, only one usage of 'event_queue' across the repo. In
this CL, the 'event_queue' argument is removed from the EventClient
classes, which should make the interface more clear. Also, the caller
is still clean and maybe more clear. The only difference is to pass a
lambda to the 'callback' argument.

BUG=none
TEST=make test

Change-Id: I9e0b8b9df542b80583ea6ed058da8738e37b221d
Reviewed-on: https://chromium-review.googlesource.com/702176
Commit-Ready: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>