pytests: Merge display_idle with display.
The display_idle pytest is not used by anyone, and the code are most
duplicated from display pytest. To avoid double effort to maintain these
two tests, we can delete the display_idle, and add an argument for
display pytest to just show an image for given duration and pass.

BUG=chromium:748540
TEST=make test

Change-Id: Ie9ca60068affaace5c91f1c034dc894f4527288e
Reviewed-on: https://chromium-review.googlesource.com/717997
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>