device: Remove interface to read dev switch status from TPM
This interface currently does not work on devices with Cr50, we're
still investigating it on chromium:748435. However, the only usage of
this interface is removed in CL:344564, so maybe it's a better idea
to just remove it (at least for now). We can added it back if we really
need to read the virtual dev switch status from TPM at one day.

BUG=chromium:748435
TEST=make test

Change-Id: I42d7d2ea4d47a028f1ebdfd3208bf53ce668e024
Reviewed-on: https://chromium-review.googlesource.com/720696
Commit-Ready: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>