audio: Allow construct alsa controller without using UCM
Add an option to force not use UCM configs. This should be a relatively
rare case, and in most of the cases, the utility function should still
be able to determine to use UCM configs or not automatically.

In case one want to force not use UCM configs, the board class need to
be modified as follows.

from cros.factory.device.audio import alsa

class MyNewBoard(chromeos.ChromeOSBoard):

  @types.DeviceProperty
  def audio(self):
    return alsa.AlsaAudioControl(self, use_ucm=False)

TEST=make test
BUG=none

Change-Id: I3a545e1eebf26a389005b0474162dade098716ec
Reviewed-on: https://chromium-review.googlesource.com/720721
Commit-Ready: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Yilun Lin <yllin@chromium.org>