test_ui: Change how static directory is resolved.
The current implementation to get the executing pytest script location
has limitation that test_ui.UI() has to be called in the pytest itself.

In order to have a better interface for test_ui, and call test_ui.UI()
in some superclass of TestCase instead, change the process of getting
where the pytest script is at by storing it in a environment variable
by invocation.

BUG=b:64963909
TEST=make test
TEST=manually on DUT, run test that have static directory (keyboard).

Change-Id: Ia16e1acaddc79dd742660a9df2df7d4142259fb9
Reviewed-on: https://chromium-review.googlesource.com/737392
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>