Instalog: Fix a failure when creating GCS client in output GCS plugin
Google Cloud Storage is depend on bucket instead of project, so we don't
need to put project name to the client. However, the client object is
general Google Cloud client, so the project name can't be None; instead
it can be an empty string.

BUG=None
TEST=manually test on local device

Change-Id: Ifcdf403bfdb58e930e40b4bac61101abf29d3a81
Reviewed-on: https://chromium-review.googlesource.com/743221
Commit-Ready: Chun-Tsen Kuo <chuntsen@chromium.org>
Tested-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>