test_ui: Warn when event handler for UI takes too much time.
Since event handler for UI is run in the event loop, if test run some
heavy load in event handler, it would block the event loop and the UI
would be irresponsible (especially the "Mark failed" button would stop
working).

Warn about this case so that developer of pytests can be aware of the
potential issue.

BUG=b:64963909
TEST=make test
TEST=manually on DUT.

Change-Id: Ib40d0774c032f09be52fba03c59c7997d4e715df
Reviewed-on: https://chromium-review.googlesource.com/742779
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>