test list: don't reset automated sequence on skipped tests
For a test list like:

  G  # AutomatedSequence, status=SKIPPED
    a  # status=PASSED
    b  # run_if='device.foo.b', STATUS=SKIPPED
    c  # status=PASSED

When we press 'PLAY' button, group 'G' will be reset because 'G' is not
'PASSED'.  However, the only test that is not passed, 'b', will still be
skipped this time.  So, the only tests will be run again are those tests
taht are already passed.  This behavior is not very intuitive.

In this CL, when the iterator is at 'G', it will peak the subtests of
'G', if all of those are SKIPPED will still be skipped this time, the
iterator will not reset the status of 'G'.

BUG=b:68919687
TEST=make test

Change-Id: I1b3bacfba76337af9c3e33ab3d46d897ec7a3ff2
Reviewed-on: https://chromium-review.googlesource.com/753299
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>