Instalog: Change the process flow of the plugins using GCS.
The process flow we are using now duplicate all events, and it slow down
the whole Instalog. To improve the performance, we change the process
flow of output_bigquery and output_archive plugins.

The process flow before this CL:
1. input -> buffer -> output_gcs -> buffer -> output_bigquery
2. input -> buffer -> output_archive -> buffer -> output_gcs

The process flow after this CL:
1. input -> buffer -> output_bigquery
2. input -> buffer -> output_archive

BUG=b:112063669, b:112015018
TEST=manually test on the local device

Change-Id: Icb47697ba9cc6864ef92dc65098d2e2fce71c202
Reviewed-on: https://chromium-review.googlesource.com/1158312
Commit-Ready: Chun-Tsen Kuo <chuntsen@chromium.org>
Tested-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>