arg_utils: Deprecate 'optional' in Arg
The 'optional' and 'default' args in Arg are redundant. If an Arg has
default value, it is indeed optional. To set an Arg as optional, provide
a default value (can be None). If an Arg does not have a default value,
it is not optional.

BUG=chromium:783704
TEST=make test

Change-Id: I3fdb975619fb9a8c18ce05cac50e310a934b3037
Reviewed-on: https://chromium-review.googlesource.com/778602
Commit-Ready: Cheng-Han Yang <chenghan@chromium.org>
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Yong Hong <yhong@google.com>