test list iterator: retest skipped test if its run_if is not set
For runnable tests (parallel test, leaf test), if it was skipped by
'skipped_test' and it is removed from 'skipped_test', it won't be
retested by current logic.

Change the way we deal with run_if so this kind of test will be
retested.

BUG=chromium:788264
TEST=make test

Change-Id: Id42f4931eb9dac0c3ba66aa5b9fd2d6037f1edd5
Reviewed-on: https://chromium-review.googlesource.com/788647
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Youcheng Syu <youcheng@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@chromium.org>