Fix missing destroy_test event when invocation is aborted.
If the invocation is aborted before the _run is started, the invocation
_run would exit immediately, and the destroy_test event won't be send by
the invocation. So the tab in frontend UI would stuck, and some
assertion would fail in JavaScript.

Fix this by do not early exit even in _aborted is set when running _run.

BUG=b:70324088
TEST=make test
TEST=manually, on remote debugging tool for Chrome of factory page, run
     'for (let i = 0; i < 1000; i++) goofy.startAutoTest();', and
     observe that there's no assertion failure after all 1000 restarts
     are completed.

Change-Id: Iab60d72faf1c5742845e2be462386e96914b2628
Reviewed-on: https://chromium-review.googlesource.com/816695
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Yong Hong <yhong@google.com>