event: Fix timeout for recv also applies to send.
The EventClientBase use same socket for both recv and send, and cause
the send to have same timeout with recv, which can cause timeout in send
when the system is in heavy load or the receive timeout is short.

Fix this by having two separate sockets to the path.

BUG=b:70537571
TEST=make test

Change-Id: Iccd8f202057a0e62599736e83624c4c0270a6875
Reviewed-on: https://chromium-review.googlesource.com/822012
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>