tablet_mode: Add argument prompt_flip_tablet.
Add an argument prompt_flip_tablet, similar to the argument in
tablet_rotation, so that this test can be used as a "begin tablet mode
tests" or "end tablet mode tests".

We can then remove the two prompt_flip_* arguments in tablet_rotation,
and use the tablet_mode test in test list similar to station_entry.

Also fix a bug that the evdev threads would be started twice because of
a bug in the python-evdev library.

BUG=None
TEST=make test
TEST=manually on DUT.

Change-Id: Id785c9b7cc9b83714b5948de5cbc93ce774ce19b
Reviewed-on: https://chromium-review.googlesource.com/831321
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>