tests: Remove 'dptf' from managed CPU frequency services list.
The 'dptf' is a daemon that will change CPU frequency to prevent
overheating.

According to SOC vendor thermal expert, they will not recommend
disabling DPTF even though CPU can do HW throttling. The thermal
solution is designed by project and they will have different needs in
order to pass the safety certification. For factory stressing, we are
not sure if its reasonable to disable it in order to make system
running in full speed. When system goes to shipping image, the system
still cant run in full speed since the DPTF is enabled.

As a result, we are removing 'dptf' from services that will be disabled
in stress testing.

BUG=b:69146283,b:70736520
TEST=make test

Change-Id: I5319ca089dbea51d55385ac1c8284d8129c01fe7
Reviewed-on: https://chromium-review.googlesource.com/832304
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Michael H Yang <michael.h.yang@intel.com>