hwid: Moves the VPD value check into `gooftool verify_vpd` sub-command.
This CL modifies `gooftool verify_vpd` sub-command so that it
verifies not only values of RO VPD but also values of RW VPD.

This CL also removes the default validation rules for VPD values from
the HWID framework to prevent non-related and duplicated rules
to be evaluated when the HWID framework is generating the HWID string.

And since in general case VPD data is not required when the HWID
framework is generating/verifying the HWID string, this CL also makes
the HWID framework default not to obtain the VPD data, the caller
should explicitly ask the HWID framework to obtain the VPD data if
necessary.

BUG=chromium:796556
TEST=make test

Change-Id: I8f27127415307dda8d05d99b1c4e50cc1a589345