pytests: Fix camera test JavaScript error handling.
The JavaScript error handling used to call test.fail directly, but since
now test.fail would only fail the test when the background test thread
ends or throws an exception (by calling WaitTaskEnd), it would deadlock
if JavaScript throws an error, and the test would hangs forever.

Fix this by manually handling the error and calling FailTask.

Note that the limit of test.fail and WaitTaskEnd would probably be
changed in a later CL, so this is a temporary workaround.

BUG=b:68681772,b:70537571
TEST=make test
TEST=manually on DUT.

Change-Id: Id3f9a3d8226877953028858e10bbbaf436235359