goofy: Deprecate 'never_fails', add 'allow_reboot' option
The 'never_fails' option basically does two things:

1. Never fail: If the status is marked as FAILED, it becomes UNTESTED.
2. Allow reboot: If DUT reboots when running this test, the test list
   will still auto run the pending tests at next boot.

After reviewing current usages of 'never_fails', we all use it for the
purpose of allowing-reboot. So this CL renames the option, and remove
the never-fail logic.

A typical usage is for a station-based test when the DUT is not attached
to a display. Note that the operator might move DUT to different test
stations to perform the test multiple times, and the test should be
considered as passed if any one of them is passed. To doing this, we
need a wait_external_test test, and set up the 'allow_reboot' and
'retries' options. So, the test can be restarted automatically at every
reboots, and pass if anyone of the test is passed.

BUG=none
TEST=make test
TEST=manually test

Change-Id: I630de176ee2018e9f71449ead241fc6b742ca80c