test_ui: Add ui.AdvanceProgress.
Since most usage of progress bar are for a fixed number of jobs that
would be done one by one, add a total number of jobs argument to
DrawProgressBar, and add an AdvanceProgress to advance the progress by
one.

Only the tests using the new TestCaseWithUI is changed, and the behavior
of ui_templates.TwoSections is not changed. (Since it would be
deprecated soon.)

BUG=b:68681772
TEST=make test
TEST=manually on DUT.

Change-Id: Ibcddd21eb696a12e6728b838f6ae80a4b05240d2