test_ui: Refactor logic of FailTask handling.
Refactor the logic of how FailTask is handled, so FailTask eventually
stop the test assuming that there's no long-running handler in event
loop, but the background thread can still be cleanly stopped by calling
PassTask / FailTask in the thread or calling WaitTaskEnd periodically.

This should not change the behavior of any test, but makes it easier to
write test that want to ignore the background thread when failing, and
should avoid issue like b:70537571.

Also makes the "Mark failed" button on screen exactly same as calling
FailTask.

BUG=b:68681772
TEST=make test
TEST=manually on DUT.

Change-Id: Id2ce1ebed57d3a6bba604d4ac3976b8b51fd87b8