audio: Separate out Initialize method from actions
In the current design, there's a special action called 'initial', to
initialize the sound cards. And in recent projects, this is the main
reason why we need the JSON/YAML audio config.

To move a step forward to make UCM config to be easily used, we separate
out the Initialize method, and UCM config can cover this step well.

Also, prevent calling volume adjugement methods by default, since it's
not supported by many recent projects, and neither by UCM configs. If
you really need to adjust the volume, you need to figure out the mixer
commands, and fill it in the audio config file.

BUG=none
TEST=make test
TEST=manually test on Pixelbook

Change-Id: I7ec3a230f7787ee7cbc04ada4b82275e5faba29b