audio: Decouple UCM config with audio config
In the current design, the UCM config manager is backed by an audio
config manager. So, if a operation is not supported by the UCM config,
the operation will be passed to audio config manager, to see if it can
be processed there.

However, after examining the real-world usage, I think this might be
an over design. Most of the usages from audio-related pytests can be
covered by UCM config itself.

In this CL, the logic of choosing UCM or audio config manager is
simplified as follows:
1. If an audio config file is provided, use it;
2. otherwise, use UCM config

So, if a UCM config is ready, no audio config file is needed at all. All
the audio-related pytests should be ready to use. For certain special
cases which need special mixer commands, a complete audio config should
be prepared, and it will switch to audio config manager automatically.

BUG=none
TEST=make test
TEST=manually test

Change-Id: I30a463858ccfc5bcfd143b7fac4b26d29a3c3f50