hwid: Decouples the probed results and the hwid v3 framework.
This is the first step of decoupling the probed results and the HWID
database.  As the `components` fields in the HWID database becomes
optional, the relation between the probed results and the HWID database
become weak.  Both the material to generate a HWID string and the result
of decoding a HWID string should be a BOM object.

TEST=make test
BUG=chromium:689944

Change-Id: I9da3778d51094a9228d8980aa1df93d474a3e09a
Reviewed-on: https://chromium-review.googlesource.com/844438
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Yilun Lin <yllin@chromium.org>