test_ui: Remove cleanup for AddTask.
Since the cleanup argument is only used in one place, and it's
equivalent to adding finally block to the task function, we remove it.

Instead, extra arguments to AddTask would be passed to the task
function, which is more useful and can save many type_utils.BindFunction
calls.

BUG=b:68681772
TEST=make test
TEST=manually on DUT.

Change-Id: I6dd944054d91322ac6ba3223d699ae092fff9664
Reviewed-on: https://chromium-review.googlesource.com/844446
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>