pytests: Remove mic_source type from audio_loop
Audio loop defines an enum to indicate which mic source should be turned
on, and is only used for its argument 'mic_source'. This mapping should
be a one-to-one mapping to the InputDevices defines in our audio device
API. However, having two enum describing the same thing makes it hard
to maintain consistency. In fact, the mic type Dmic2 is missing from
mic_source, which should be a bug.

In our last code refactoring in audio device API, we defined the enum
InputDevices in API, and implement a new EnableDevice API. The enum
'mic_source' defined in audio_loop pytest should no longer necessary.

After this change, the mic_source should be revised as follows:
1. 'external' --> 'Extmic'
2. 'panel' --> 'Dmic'
3. 'mlb' --> 'MLBDmic'

BUG=none
TEST=make test
TEST=manually test

Change-Id: Id4bab33e7dc919441f706a72ad336e0990d329ee
Reviewed-on: https://chromium-review.googlesource.com/844495
Commit-Ready: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Shen-En Shih <petershih@chromium.org>