appengine: Porting HWID API from google3.
To reuse the APIary config, which exposes AppEngine API to
googleapis.com, such that both IP from interior or exterior
of google domain can access the API. We decide to re-use the
APIary config by porting the HWID API from google3. The source
can be found in go/g3hwidapi. And also, the coding style is very
different between two places, we have to modify a bit. In addition,
the available python packages are different, too, so we have to
replace them with the self-implemented one.

- Follows factory coding style.
- Replaces google3 package pyglib to class _RamSize
- Adapts hwid_manager to new HWID v3 API, including
  - Remove unnecessary file hwid_yaml_loader.py, security.py.
- Renames board names in testdata.
- Adds AppEngine integration test.
- Adds deploy script.
- Sets app.yaml instance_class to F4_1G, which is the fastest and the largest
  memory machine.

BUG=chromium:818125
TEST=manual test; make test; appengine/test/integration_test.py

Change-Id: I970eadaebe9c8b8e2c222b891c9966fe3b42f2ca
Reviewed-on: https://chromium-review.googlesource.com/963879
Commit-Ready: Yilun Lin <yllin@chromium.org>
Tested-by: Yilun Lin <yllin@chromium.org>
Reviewed-by: Yong Hong <yhong@google.com>