instalog: Fix race condition in buffer_simple_file.
When multiple threads call Produce, the tmp_dir may be deleted by some
thread when other is still inside the function. This cause the ending
shutil.rmtree to throw an exception.

Change the tmp_dir to use TempDirectory instead of a fixed name to fix
this.

BUG=None
TEST=make test

Change-Id: I99040619b7e408def9f4a4d00fff6c8a9ca68acc
Reviewed-on: https://chromium-review.googlesource.com/850012
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>