test_ui: Remove DummyUI.
Remove DummyUI since no code is currently using it, and we won't need it
in the future.

If Goofy need to be run without an UI, not starting the websocket to UI
in Goofy should be enough, and no change in test_ui should be needed.
Since all test_ui module do is send events to the event server and do
NOT directly interact with Chrome.

BUG=b:68681772
TEST=make test

Change-Id: If031e450ff7fe2ab082a46f4842157975600aa4d
Reviewed-on: https://chromium-review.googlesource.com/853833
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>