sys_utils: Add a function IsContainerized.
Previously umpire_env checks if inside docker by comparing the first
token in /proc/1/sched with string 'init'. This is obviously not a good
way (and on some new OS this value might be 'systemd').

According to https://docs.docker.com/engine/admin/runmetrics/, we can
look at /proc/1/cgroup and see if we can currently inside docker cgroup.

Due to the reason that we might have more docker containers in factory
repository in the future, we should move this function to sys_utils so
that we can re-use this function in other containers.

BUG=none
TEST=manually

Change-Id: Id2a72890425dffacf3d1a52bff8f34041671a046
Reviewed-on: https://chromium-review.googlesource.com/867184
Commit-Ready: Youcheng Syu <youcheng@chromium.org>
Tested-by: Youcheng Syu <youcheng@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>