default_test_ui: Show resolved argument.
Currently the default_test_ui only shows original argument as is in JSON
test list. It may contain things like "eval! " and "i18n! ", which is
inconvenient when debugging. Default to show the resolved argument
instead, and add a toggle to show the original argument as it is in JSON
test list.

BUG=b:68681772
TEST=make test
TEST=manually on DUT.

Change-Id: Ia60dc1bf9d24c91c62de04e03e164a6a9dab09d2
Reviewed-on: https://chromium-review.googlesource.com/877744
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>