goofy Makefile: Rebuild closure.css when Makefile changed.
Rebuild the closure.css when Makefile is changed, since the dependency
list CSS_SOURCE_PATHS is located in Makefile, and currently if someone
change the list, the closure.css would not be rebuilt.

This would cause some extra rebuild if there's unrelated change to
Makefile, but since the Makefile is rarely changed, and the build time
is almost instantaneous, this should not cause too much extra work.

BUG=None
TEST=make test
TEST=manually.

Change-Id: I3b434e9dd6a87ecf7be9c3fdf4ededd0313ac1f3
Reviewed-on: https://chromium-review.googlesource.com/877206
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>