hwid: Rewrites `DatabaseBuilder` class and its unittest.
In previous CLs we implemented method for manipulating a HWID database
object in `Database` class and we also removed some complex logics
such as "probeable" from the database.  It's time to rewrite the
builder for the database in a simpler way.

This CL also moves the probing related functions from `hwid_utils`
module into a individual module to focus on "how to use the probe
tool to know what components are installed on the device".

TEST=make test; manually test all functions of hwid tools on DUT
BUG=chromium:689944
CQ-DEPEND=CL:*553738,CL:*553758

Change-Id: I2e79298ff49b304b9c17be1dfa21e95ec8c12232
Reviewed-on: https://chromium-review.googlesource.com/885343
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>