inform_shopfloor: Explicitly return 0 on informing success.
TTY for devices without a connected monitor/display may disappear on
wiping staging, thus return non-zero value when we pipes to the TTY
device. Since the goal to pipe information to TTY is to show the
current status on screen, we can just expliclty return True for the case.

TEST=run finalize without a connected monitor, and check it shutdown.
BUG=chromium:804762

Change-Id: I74fbc9a80bf5e91e30b303bf94ae2ed2abd5189a
Reviewed-on: https://chromium-review.googlesource.com/895215
Commit-Ready: Yilun Lin <yllin@chromium.org>
Tested-by: Yilun Lin <yllin@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>