probe: Add an new field "information" in the probe statement.
This CL revises the syntax of the probe statement and the probed
results to let the user be able to add additional information in
the probe statement and let the probe tool will forward that
information to the probed results.

To deprecate the "components" part of HWID databases, we have to find
another place, which must be also included in HWID bundles, to record
the status of each component.  As the source of the probe statement
will eventually become the AVL, it's reasonable to record the
component status in the probe statement.

BUG=chromium:816370
TEST=make test

Change-Id: Ieaa4cbedd974d717f7dfdd46e2490d7d1f1b32d1
Reviewed-on: https://chromium-review.googlesource.com/938867
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Yilun Lin <yllin@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>