pylint: Fix "Redefining built-in 'exit'" on "make lint"
The pylint has weird issue that when checking the file
py/test/__init__.py, it tries to import py.test. But in chroot,
there's a standard library py.test that is "<AliasModule 'py.test' for
'pytest'>", and pylint resolve the module py.test incorrectly and found
the stdlib one, which cause error on py.test.exit redefining the
built-in exit.

Fix it by adding the empty file py/test/__init__.py to the blacklist for
linting.

BUG=b:73868308
TEST=make test
TEST=make lint, and see that there's no weird "Redefining built-in
     'exit'" warning with no file associated.

Change-Id: I163f0d2ed51649b6fcac283e42c790ecf1b61863
Reviewed-on: https://chromium-review.googlesource.com/948383
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>