event_log_unittest: Fix flaky test.
When the system is on heavy disk load (for example, when running "make
test" and other disk-heavy test is run in parallel), the initialization
of event_log.GlobalSeq can take a long time in the os.fsync() call,
causing the number of events less than expected.

Mock the os.fsync call in unittest to fix this.

BUG=None
TEST=make test
TEST=run the test when the disk is on heavy load. (For example, running
     `stress -d 8` in parallel)

Change-Id: I3790294f2fea62010f2424d2b686ef146cc7d129
Reviewed-on: https://chromium-review.googlesource.com/958750
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>