init: Change /var/run to be symlink for wipe-in-place.
ToT test images have changed /var/run from symlink to `mount --bind`.
However, this will break wipe-in-place flow due to not able to umount
/var/run (locked by dhcpcd).

It is possible to change dhcpcd, but we may see more issues in future.
So the solution here is to re-create /var/run as symlink when factory
toolkit (goofy device) is enabled.

BUG=b:74420122
TEST=None

Change-Id: Id29e85bfa25ecc03d6362b0fe1c48c04ad1915b6
Reviewed-on: https://chromium-review.googlesource.com/958784
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Shen-En Shih <petershih@chromium.org>